cortex-m: move shared code to cortexm crate
